Initial Commit
This includes scripts to decompile Minecraft and setup a workspace.
Building the Maven project will recompile Minecraft with the applied
patches.

There are a number of decompiler errors that need to be patched.
Included are a number of patches that fix the first few, but there are
still many more to fix.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>