Switch to Forge's FernFlower and use MCP's mappings
Forge's FernFlower fork is a lot more stable than Spigot's, and MCP's
mappings are the best we have. FernFlower still isn't perfect, so there
are a lot of errors to fix. Most of the have to do with generics.

Mappings are from https://github.com/Techcable/MinecraftMappings
Fernflower is from https://github.com/MinecraftForge/FernFlower

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>